feat #21 :truck: move package to core.ui.lifecycle
