+++ presentation/src/main/java/com/dcs/presentation/ui/setting/SettingViewModel.kt
@@ -9,7 +9,7 @@ import com.dcs.presentation.core.model.SettingUiState
 import com.dcs.presentation.core.model.UserProfile
 import com.dcs.presentation.core.state.UiState
 import com.dcs.presentation.core.state.asUiState
-import com.dcs.presentation.core.ui.launch
+import com.dcs.presentation.core.ui.lifecycle.launch
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
